Optimize database inserts/deletes
Previously we went through all post updates and added them or deleted them one at a time.
This was inefficient when there were two post updates that created and deleted a post.
Now we add the posts to the database after we've created the comment collection, so that only posts without corresponding deletes are added.