Optimizing UI updates
Iterate over the comments and search for the post.

This will be faster because we can use a binary search for the posts,
and the number of comments retrieved will be small relative the posts that we have in our timeline.

TODO Need to update the search method to use binary search.