Check post user id when invoked
This fixes issue #61

When the feed changed on refresh, the display changed for the new posts, but delete was still bound to the post holders at positions where my posts were.

If my post was in position 2, when a new post came in, there would still be a delete option on the post in postition 2, whether or not it was my post.

The code now checks user ID when the listener is invoked. Since it checks at "run time" it uses the information for the post that is currently there